+++ dubbo-samples-group/src/main/java/org/apache/dubbo/samples/group/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.group;
+package org.apache.dubbo.samples.group;
 
 import java.io.File;
 import java.lang.reflect.Method;
